Use size_t instead of int for anything that's a buffer size or index

+++ include/Aulib/AudioDecoder.h
@@ -38,7 +38,7 @@ public:
     static AudioDecoder* decoderFor(SDL_RWops* rwops);
 
     bool isOpen() const;
-    int decode(float buf[], int len, bool& callAgain);
+    size_t decode(float buf[], size_t len, bool& callAgain);
 
     virtual bool open(SDL_RWops* rwops) = 0;
     virtual unsigned getChannels() const = 0;
@@ -50,7 +50,7 @@ public:
 protected:
     void setIsOpen(bool f);
 
-    virtual int doDecoding(float buf[], int len, bool& callAgain) = 0;
+    virtual size_t doDecoding(float buf[], size_t len, bool& callAgain) = 0;
 
 private:
     AudioDecoder(const AudioDecoder&);
