+++ include/Aulib/AudioDecoderVorbis.h
@@ -34,7 +34,7 @@ public:
     bool open(SDL_RWops* rwops) override;
     unsigned getChannels() const override;
     unsigned getRate() const override;
-    int doDecoding(float buf[], int len, bool& callAgain) override;
+    size_t doDecoding(float buf[], size_t len, bool& callAgain) override;
     bool rewind() override;
     float duration() const override;
     bool seekToTime(float seconds);
