+++ src/AudioResamplerSpeex.cpp
@@ -65,7 +65,7 @@ Aulib::AudioResamplerSpeex::~AudioResamplerSpeex()
 
 
 void
-Aulib::AudioResamplerSpeex::doResampling(float dst[], const float src[], int& dstLen, int& srcLen)
+Aulib::AudioResamplerSpeex::doResampling(float dst[], const float src[], size_t& dstLen, size_t& srcLen)
 {
     if (d->fResampler == nullptr) {
         dstLen = srcLen = 0;
